Moving the llm models from the example folder to p-kit/llm
