+++ examples/llm/llm_demo_003.py
@@ -29,7 +29,7 @@ import numpy as np
 import torch
 from torch.utils.data import DataLoader
 from transformers import GPT2Config, GPT2LMHeadModel
-from llm_model import SparsePBitLMTemporalMemory
+from p_kit.llm.llm_models import SparsePBitLMTemporalMemory
 
 TRAIN = """
 the cat sat on the mat.
