+++ p_kit/llm/__init__.py
@@ -1,0 +1,6 @@
+from .llm_models import SparsePBitLM, SparsePBitLMTemporalMemory
+
+__all__ = [
+    "SparsePBitLM",
+    "SparsePBitLMTemporalMemory",
+]
